+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - Fix syntax color for odoc math blocks containing braces (#1741)
 
+- Add one-click install option for ocaml-lsp-server (#1725)
+
 ## 1.28.0
 
 - Add `.re` file extension support. (#1685)